Preserve ZeroMQ SDK status retry metadata
